+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.4.6
 - WELD folded plane
+- Add debug button to display overlapping faces
 - Box build with 0 height and abs(depth) > 0 creates a plane (not a thin box)
 - Undoing does not clear previously selected image
 - Entering vertex paint enables face orientation (and leaving vertex paint mode returns to the previous value)